Fixed RefreshData() and added PintData()
Refreshing now gets done before sorting to make sure the data is still accurate and up to date with the xml file. After sorting it gets placed in the table with printdata.

The refreshdata button now also works and resets the sorting of the data to the way its sorted in the xml document, resetting it to its original state.